fix(cli): add waitForSync to ensure proper store synchronization
This change adds a waitForSync function that waits for the store to finish syncing before shutdown to prevent data loss. The function is used in two places:
1. Before shutting down the store in the main CLI exit handler
2. In the preAction hook to ensure sync before command execution

The waitForSync function takes an optional store parameter and a timeout duration, defaulting to 1 second.
